llama_index_agents/pyproject.toml
[project]
name = "llama-index-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.5",
    "browser-use>=0.1.40",
    "bs4>=0.0.2",
    "llama-index-llms-openrouter>=0.3.2",
    "llama-index>=0.12.40",
    "playwright>=1.52.0",
    "playwright-stealth>=1.0.6",
    "python-dotenv>=1.1.0",
    "selenium>=4.33.0",
    "setuptools>=80.9.0",
    "matplotlib>=3.10.3",
    "together>=1.5.8",
    "loguru>=0.7.3",
    "asyncclick>=8.1.8",
    "jwt>=1.3.1",
    "pyjwt>=2.10.1",
]
